+++ plugin-server/src/worker/vm/upgrades/export-events.ts
@@ -1,9 +1,9 @@
 import { Plugin, PluginEvent, PluginMeta, RetryError } from '@posthog/plugin-scaffold'
 
 import { Hub, MetricMathOperations, PluginConfig, PluginConfigVMInternalResponse, PluginTaskType } from '../../../types'
+import { determineNodeEnv, NodeEnv } from '../../../utils/env-utils'
 import { status } from '../../../utils/status'
-import { determineNodeEnv, stringClamp } from '../../../utils/utils'
-import { NodeEnv } from './../../../utils/utils'
+import { stringClamp } from '../../../utils/utils'
 import { ExportEventsBuffer } from './utils/export-events-buffer'
 
 export const MAXIMUM_RETRIES = 10
